pyproject.toml
[tool.poetry]
name = "password-manager"
version = "0.1.0"
description = ""
authors = ["Mateusz <morveusz93@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pyperclip = "^1.8.2"
Pillow = "^9.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
